requirements.txt
chromadb==1.3.4
langchain==0.1.0
langchain-core==1.0.5
langchain-community==0.2.19
langchain-chroma==0.6.1
langchain-text-splitters==1.0.0
langchain-huggingface>=0.0.4
sentence-transformers==5.1.2
pypdf==6.2.0
python-dotenv==1.2.1
